bisa getbyid

+++ model/albums/albums.go
@@ -1,9 +1,9 @@
 package album
 
 type Album struct {
-	AlbumId       int     `json:"id"`
-	Title         string  `json:"title"`
-	Artist_Id     int  `json:"artist_id"`
-	Price         float32 `json:"price"`
-	Year int `json:"year"`
+	AlbumId   int     `json:"id"`
+	Title     string  `json:"title"`
+	Artist_Id int     `json:"artist_id"`
+	Price     float64 `json:"price"`
+	Year      int     `json:"year"`
 }
